refactor: move send tweets to helper folder
