+++ user-service/src/handlers/subscribe.ts
@@ -1,11 +1,9 @@
 import { APIGatewayProxyResult } from 'aws-lambda'
 import { commonMiddleware } from '../lib/middlewares/commonMiddleware'
 import { SubscribeService } from './services/subscribeService'
-import { SESMailProvider } from '../lib/providers/implementations/SESMailProvider'
 import { DynamoRepository } from '../lib/repositories/implementations/DynamoRepository'
 
 const service = new SubscribeService(
-  new SESMailProvider(),
   new DynamoRepository()
 )
